+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 v1.0.2
+- templateID 更名为 templateFile
 - 修改示例数据
 - 完善文档,注释